prediction/requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
psycopg2==2.8.6
# numpy==1.18.5
# transformers==3.5.1
# gunicorn==19.7.1

# torch==1.7.0
# torchvision==0.8.1
# torchaudio==0.7.0

pymystem3==0.2.0
scipy==1.6.1
sklearn==0.0
numpy==1.18.5
pandas==1.2.2
